+++ src/main/kotlin/lean4ij/actions/ToggleLeanInfoViewInternal.kt
@@ -2,6 +2,7 @@ package lean4ij.actions
 
 import com.intellij.icons.AllIcons
 import com.intellij.ide.BrowserUtil
+import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -18,6 +19,7 @@ import lean4ij.infoview.LeanInfoViewWindowFactory
 import lean4ij.infoview.LeanInfoviewService
 import lean4ij.infoview.external.JcefInfoviewService
 import lean4ij.infoview.external.JcefInfoviewTooWindowFactory
+import lean4ij.setting.Lean4Settings
 import lean4ij.util.LeanUtil
 import lean4ij.util.notify
 
@@ -55,15 +57,6 @@ class ToggleLeanInfoViewJcef : AnAction() {
             toolWindow.hide()
         } else {
             toolWindow.show()
-        }
-    }
-
-    override fun update(e: AnActionEvent) {
-        val editor = e.dataContext.getData<Editor>(CommonDataKeys.EDITOR)?:return
-        // TODO here it can be null, weird
-        val virtualFile = editor.virtualFile?: return
-        if (!LeanUtil.isLeanFile(editor.virtualFile)) {
-            e.presentation.isVisible = false
         }
     }
 
@@ -215,6 +208,35 @@ class ToggleInternalInfoviewSoftWrap : AbstractToggleUseSoftWrapsAction(SoftWrap
             } catch (ex: TimeoutCancellationException) {
                 null
             }
+        }
+    }
+}
+
+class ToggleInfoviewPreferred : AnAction() {
+    private val lean4Settings = service<Lean4Settings>()
+    // TODO constant for the actions
+    private val toggleLeanInfoViewInternal = ActionManager.getInstance().getAction("ToggleLeanInfoViewInternal")
+    private val toggleLeanInfoViewJcef = ActionManager.getInstance().getAction("ToggleLeanInfoViewJcef")
+
+    override fun getActionUpdateThread(): ActionUpdateThread {
+        return ActionUpdateThread.BGT;
+    }
+
+    override fun actionPerformed(e: AnActionEvent) {
+        when {
+            lean4Settings.preferredInfoview == "Jcef" ->
+                    toggleLeanInfoViewJcef.actionPerformed(e)
+            else ->
+                    toggleLeanInfoViewInternal.actionPerformed(e)
+        }
+    }
+
+    override fun update(e: AnActionEvent) {
+        val editor = e.dataContext.getData(CommonDataKeys.EDITOR)?:return
+        // TODO here it can be null, weird
+        val virtualFile = editor.virtualFile?: return
+        if (!LeanUtil.isLeanFile(virtualFile)) {
+            e.presentation.isVisible = false
         }
     }
 }